+++ lib/extensions/mod.rs
@@ -3,3 +3,6 @@ pub mod ratelimit;
 
 #[cfg(feature = "logging")]
 pub mod logger;
+
+#[cfg(feature = "serve_static")]
+pub mod serve_static;