+++ Changelog.md
@@ -1,5 +1,7 @@
 # 0.2.1*
 - Only Build common::remove_address_port if logger or rate-limiter are enabled
+- Make Header name / value Public
+- Serve Static Middleware
 
 # 0.2.0
 - Response Overhaul, Now more like a Response Builder